code/requirements.txt
###### Requirements without Version Specifiers ######
pyyaml
pymysql
flask
flasgger
